add pyreneewHEWParam dataclass
