+++ pipelines/forecast_pyrenew.py
@@ -14,13 +14,12 @@ from fit_pyrenew_model import fit_and_save_model
 from generate_predictive import (
     generate_and_save_predictions,
 )
-from prep_data import process_and_save_loc
+from prep_data import process_and_save_loc_data, process_and_save_loc_param
 from prep_eval_data import save_eval_data
 from prep_ww_data import clean_nwss_data, preprocess_ww_data
 from pygit2.repository import Repository
 
 from pyrenew_hew.utils import (
-    approx_lognorm,
     flags_from_hew_letters,
     pyrenew_model_name_from_flags,
 )
@@ -175,141 +174,6 @@ def create_hubverse_table(model_fit_path):
     if result.returncode != 0:
         raise RuntimeError(f"create_hubverse_table: {result.stderr}")
     return None
-
-
-def _validate_and_extract(
-    df: pl.DataFrame,
-    parameter_name: str,
-    allow_missing_right_truncation: bool = False,
-) -> list:
-    if (
-        allow_missing_right_truncation
-        and parameter_name == "right_truncation"
-        and df.height == 0
-    ):
-        return list([1])
-    if df.height != 1:
-        error_msg = f"Expected exactly one {parameter_name} parameter row, but found {df.height}"
-        logging.error(error_msg)
-        if df.height > 0:
-            logging.error(f"Found rows: {df}")
-        raise ValueError(error_msg)
-    return df.item(0, "value").to_list()
-
-
-def get_pmfs(
-    param_estimates: pl.LazyFrame,
-    loc_abb: str,
-    disease: str,
-    as_of: dt.date = None,
-    reference_date: dt.date = None,
-    allow_missing_right_truncation: bool = False,
-) -> tuple[list, list, list]:
-    """
-    Filter and extract probability mass functions (PMFs) from
-    parameter estimates LazyFrame based on location, disease
-    and date filters.
-
-    This function queries a LazyFrame containing epidemiological
-    parameters and returns three types of PMF parameters:
-    delay, generation interval, and right truncation.
-
-    Parameters
-    ----------
-    param_estimates: pl.LazyFrame
-        A LazyFrame containing parameter data with columns
-        including 'disease', 'parameter', 'value', 'geo_value',
-        'start_date', 'end_date', and 'reference_date'.
-
-    loc_abb : str
-        Location abbreviation (geo_value) to filter
-        right truncation parameters.
-
-    disease : str
-        Name of the disease.
-
-    as_of : datetime.date, optional
-        Date for which parameters must be valid
-        (start_date <= as_of <= end_date). Defaults
-        to the most recent estimates.
-
-    reference_date : datetime.date, optional
-        The reference date for right truncation estimates.
-        Defaults to as_of value. Selects the most recent estimate
-        with reference_date <= this value.
-
-    allow_missing_right_truncation : bool, optional
-        If true, allows extraction of other pmfs if
-        right_truncation estimate is missing
-
-    Returns
-    -------
-    tuple[list, list, list]
-        A tuple containing three arrays:
-        - generation_interval_pmf: Generation interval distribution
-        - delay_pmf: Delay distribution
-        - right_truncation_pmf: Right truncation distribution
-
-    Raises
-    ------
-    ValueError
-        If exactly one row is not found for any of the required parameters.
-
-    Notes
-    -----
-    The function applies specific filtering logic for each parameter type:
-    - For delay and generation_interval: filters by disease,
-      parameter name, and validity date range.
-    - For right_truncation: additionally filters by location.
-    """
-    min_as_of = dt.date(1000, 1, 1)
-    max_as_of = dt.date(3000, 1, 1)
-    as_of = as_of or max_as_of
-    reference_date = reference_date or as_of
-
-    filtered_estimates = (
-        param_estimates.with_columns(
-            pl.col("start_date").fill_null(min_as_of),
-            pl.col("end_date").fill_null(max_as_of),
-        )
-        .filter(pl.col("disease") == disease)
-        .filter(
-            pl.col("start_date") <= as_of,
-            pl.col("end_date") >= as_of,
-        )
-    )
-
-    generation_interval_df = filtered_estimates.filter(
-        pl.col("parameter") == "generation_interval"
-    ).collect()
-
-    generation_interval_pmf = _validate_and_extract(
-        generation_interval_df, "generation_interval"
-    )
-
-    delay_df = filtered_estimates.filter(
-        pl.col("parameter") == "delay"
-    ).collect()
-    delay_pmf = _validate_and_extract(delay_df, "delay")
-
-    # ensure 0 first entry; we do not model the possibility
-    # of a zero infection-to-recorded-admission delay in Pyrenew-HEW
-    delay_pmf[0] = 0.0
-    delay_pmf = jnp.array(delay_pmf)
-    delay_pmf = delay_pmf / delay_pmf.sum()
-    delay_pmf = delay_pmf.tolist()
-
-    right_truncation_df = (
-        filtered_estimates.filter(pl.col("geo_value") == loc_abb)
-        .filter(pl.col("parameter") == "right_truncation")
-        .filter(pl.col("reference_date") == pl.col("reference_date").max())
-        .collect()
-    )
-    right_truncation_pmf = _validate_and_extract(
-        right_truncation_df, "right_truncation", allow_missing_right_truncation
-    )
-
-    return (generation_interval_pmf, delay_pmf, right_truncation_pmf)
 
 
 def main(
@@ -502,22 +366,6 @@ def main(
         loc_level_nwss_data = None
 
     param_estimates = pl.scan_parquet(Path(param_data_dir, "prod.parquet"))
-    (generation_interval_pmf, delay_pmf, right_truncation_pmf) = get_pmfs(
-        param_estimates=param_estimates,
-        loc_abb=loc,
-        disease=disease,
-        allow_missing_right_truncation=not fit_ed_visits,
-    )
-
-    inf_to_hosp_admit_lognormal_loc, inf_to_hosp_admit_lognormal_scale = (
-        approx_lognorm(
-            jnp.array(delay_pmf)[1:],  # only fit the non-zero delays
-            loc_guess=0,
-            scale_guess=0.5,
-        )
-    )
-
-    inf_to_hosp_admit_pmf = delay_pmf
 
     model_batch_dir_name = (
         f"{disease.lower()}_r_{report_date}_f_"
@@ -550,7 +398,7 @@ def main(
     copy_and_record_priors(priors_path, model_run_dir)
 
     logger.info(f"Processing {loc}")
-    process_and_save_loc(
+    process_and_save_loc_data(
         loc_abb=loc,
         disease=disease,
         facility_level_nssp_data=facility_level_nssp_data,
@@ -563,6 +411,14 @@ def main(
         logger=logger,
         credentials_dict=credentials_dict,
         nhsn_data_path=nhsn_data_path,
+    )
+    process_and_save_loc_param(
+        loc_abb=loc,
+        disease=disease,
+        loc_level_nwss_data=loc_level_nwss_data,
+        param_estimates=param_estimates,
+        fit_ed_visits=fit_ed_visits,
+        model_run_dir=model_run_dir,
     )
     logger.info("Getting eval data...")
     if eval_data_path is None:
@@ -595,11 +451,6 @@ def main(
         fit_ed_visits=fit_ed_visits,
         fit_hospital_admissions=fit_hospital_admissions,
         fit_wastewater=fit_wastewater,
-        generation_interval_pmf=generation_interval_pmf,
-        right_truncation_pmf=right_truncation_pmf,
-        inf_to_hosp_admit_lognormal_loc=inf_to_hosp_admit_lognormal_loc,
-        inf_to_hosp_admit_lognormal_scale=inf_to_hosp_admit_lognormal_scale,
-        inf_to_hosp_admit_pmf=inf_to_hosp_admit_pmf,
     )
     logger.info("Model fitting complete")
 
@@ -610,11 +461,6 @@ def main(
         model_run_dir,
         pyrenew_model_name,
         n_days_past_last_training,
-        generation_interval_pmf=generation_interval_pmf,
-        right_truncation_pmf=right_truncation_pmf,
-        inf_to_hosp_admit_lognormal_loc=inf_to_hosp_admit_lognormal_loc,
-        inf_to_hosp_admit_lognormal_scale=inf_to_hosp_admit_lognormal_scale,
-        inf_to_hosp_admit_pmf=inf_to_hosp_admit_pmf,
         predict_ed_visits=forecast_ed_visits,
         predict_hospital_admissions=forecast_hospital_admissions,
         predict_wastewater=forecast_wastewater,
